+++ src/main/java/se/kth/iv1201/project/domain/CompetenceProfileDTO.java
@@ -1,7 +1,5 @@
 package se.kth.iv1201.project.domain;
 
-import java.math.BigDecimal;
-
 public interface CompetenceProfileDTO {
     int getCompetenceProfileID();
     
@@ -9,5 +7,5 @@ public interface CompetenceProfileDTO {
 
     int getCompetenceID();
 
-    BigDecimal getYearOfExperience();
+    int getYearOfExperience();
 }
